third_party/blink/web_tests/platform/generic/fast/css/content-language-multiple-expected.txt

Test for bug 76701: map HTTP-EQUIV content-language to -webkit-locale. This particular test tests that in the event of multiple content-languages, the final content-language is used as the pragma-set default language. This expectation may change, see bug. The HTML5 spec decrees that the first successfully processed one be used. Firefox and IE seem to use the final one.

PASS languageOfNode('x') is "\"zh_CN\""
PASS languageOfNode('y') is "\"ar\""
PASS successfullyParsed is true

TEST COMPLETE
